pybackend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
newspaper3k
requests
beautifulsoup4
python-multipart
